// src/services/fatima/student-package-fatima/types.ts
export declare namespace NsFatimaStudentService {
    export interface AddStudentPackageCourse {
        student_id: string;
        start_at: string;
        end_at: string;
        course_id: string;
    }

    export interface UpdateStudentPackageCourse {
        student_package_id: string;
        start_at: string;
        end_at: string;
    }
}